+++ CHANGELOG.md
@@ -5,12 +5,13 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 
-## WIP
+## 0.6.0 - 6. August 2020
 ### Added
 - `[compilers]` Added copy compiler for if no compiler is found for files.
 
 ### Fixed
 - `[dev-server]` Fixed issue with double compiling files.
+- `[watch]` Fixed issue with double compiling files.
 
 ## 0.5.6 - 6. August 2020
 ### Added